explore capabilities of solsa allowing functions to be exported
